% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal.R
\name{diagonal}
\alias{diagonal}
\title{Differentiate between a diamond and a square}
\usage{
diagonal(x, y)
}
\description{
The two have something in common: the four sides are the same.
Distinguish by the length of two diagonally.
}
